refactor: use non-relative module imports for React components
As long as imports are somewhere in /app, they can refer to any module
directly in /app/components with its simple name (no need for path).
