+++ app/app.css
@@ -28,7 +28,7 @@
 @import './components/links/index.css';
 @import './components/difference/index.css';
 @import './components/logo-loader/index.css';
-@import './components/progressbar/index.css';
+@import './components/ProgressBar/index.css';
 @import './components/modal/index.css';
 @import './components/toggle/index.css';
 @import './components/switch/index.css';
